# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dedicated API for ArcheryHub Mobile App (Archer \u0026 Scorekeeper)",
        "title": "ArcheryHub Mobile API",
        "termsOfService": "http://archeryhub.id/terms/",
        "contact": {
            "name": "ArcheryHub Support",
            "url": "https://archeryhub.id",
            "email": "support@archeryhub.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1"
    },
    "host": "api.archeryhub.id",
    "basePath": "/api/v1",
    "paths": {
        "/events/{id}/results/elimination": {
            "get": {
                "description": "Get the elimination match brackets and results for a specific category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Results"
                ],
                "summary": "Get Elimination Bracket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category UUID",
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/events/{id}/results/qualification": {
            "get": {
                "description": "Get complete leaderboard for a specific category, including end-by-end scores",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Results"
                ],
                "summary": "Get Qualification Results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category UUID",
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}